process
I was facing issues with the checkout process. The prices aren't showing even the quantity in the checkout page. I believe that I have written the right code. I have added that when the user doesn't have a promo code they can still checkout. I have also added that if the user writes a promo code that doesn't exist it will show an error. For the diagnostic testing , I haven't completed it since it is difficult to make and I have been having a lot of issues with it so I removed it. I have fixed the favourites page. I changed the place of the heart icon and changed the layout of the whole page. It functions better now. Because I had duplicated images and when triggering the heart they were both triggered but I fixed it.