Add files via upload
I have fixed the same issue that I had for the favourites page for the newin and trending page. I have an issue with the home page where I have the 4 section of face care, body and hair care, stage of skin ageing and by concern. Unfortunately when I click on the face care image the body care is showing and the same for stage of skin ageing and by concern. I believe I have the right code. I have been trying to make the recent orders show in the account page in my orders section.