Updates
removed the document as it was the wrong pdf.

added a readme.md file as on github repository it won't let me update the old one and when I created a new file on the github repository it's not showing so i added it to vsc.

I updated the payment method page as the payment wasn't saving as a default. now it's working